Move system-related commands to eth_system.go
